a couple more

+++ src/platform_impl/linux/x11/util/geometry.rs
@@ -130,12 +130,12 @@ impl FrameExtentsHeuristic {
                 self.frame_extents
                     .left
                     .saturating_add(self.frame_extents.right),
-            ),
+            ) as u32,
             height.saturating_add(
                 self.frame_extents
                     .top
                     .saturating_add(self.frame_extents.bottom),
-            ),
+            ) as u32,
         )
     }
 }
